fix(linalg): CSR row_offsets to size_t + BL max_iter use n not m
1. CSR row_offsets_: uint32_t → size_t to prevent silent overflow
   when total NNZ exceeds 4.29G (latent bug for large/dense matrices)

2. BL max_iter: use n/64+100 (columns) instead of m/64+100 (rows).
   Montgomery BL converges in at most n/64 iterations. For 50-digit
   GNFS (168K rows × 53K cols), this reduces max_iter from 2725 to
   928, saving ~1800 unnecessary iterations when BL doesn't converge.

+++ include/gnfs/linalg/sparse_matrix.hpp
@@ -472,7 +472,7 @@ public:
         size_t total_nnz = 0;
         for (size_t i = 0; i < num_rows_; ++i) {
             total_nnz += mat.row(i).indices().size();
-            row_offsets_[i + 1] = static_cast<uint32_t>(total_nnz);
+            row_offsets_[i + 1] = total_nnz;
         }
 
         // Pack all column indices into one contiguous array
@@ -502,11 +502,11 @@ public:
 
     /// Access underlying data for direct iteration
     [[nodiscard]] const std::vector<uint32_t>& col_indices() const noexcept { return col_indices_; }
-    [[nodiscard]] const std::vector<uint32_t>& row_offsets() const noexcept { return row_offsets_; }
+    [[nodiscard]] const std::vector<size_t>& row_offsets() const noexcept { return row_offsets_; }
 
 private:
     std::vector<uint32_t> col_indices_;   // All column indices, packed contiguously
-    std::vector<uint32_t> row_offsets_;   // row_offsets_[i] = start of row i in col_indices_
+    std::vector<size_t> row_offsets_;   // row_offsets_[i] = start of row i in col_indices_
     size_t num_rows_ = 0;
     size_t num_cols_ = 0;
 };
